listener: split class ListenerImpl from listener_manager_impl.cc (#8596)
