+++ source/server/BUILD
@@ -257,9 +257,45 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
-    name = "listener_manager_lib",
-    srcs = ["listener_manager_impl.cc"],
-    hdrs = ["listener_manager_impl.h"],
+    name = "listener_lib",
+    srcs = [
+        "listener_impl.cc",
+    ],
+    hdrs = [
+        "listener_impl.h",
+    ],
+    deps = [
+        ":configuration_lib",
+        ":drain_manager_lib",
+        ":filter_chain_manager_lib",
+        ":listener_manager_impl",
+        ":transport_socket_config_lib",
+        ":well_known_names_lib",
+        "//include/envoy/server:active_udp_listener_config_interface",
+        "//include/envoy/server:filter_config_interface",
+        "//include/envoy/server:listener_manager_interface",
+        "//include/envoy/server:transport_socket_config_interface",
+        "//source/common/config:utility_lib",
+        "//source/common/init:manager_lib",
+        "//source/common/network:connection_balancer_lib",
+        "//source/common/network:listen_socket_lib",
+        "//source/common/network:socket_option_factory_lib",
+        "//source/common/network:utility_lib",
+        "//source/common/protobuf:utility_lib",
+        "//source/extensions/filters/listener:well_known_names",
+        "//source/extensions/transport_sockets:well_known_names",
+    ],
+)
+
+envoy_cc_library(
+    name = "listener_manager_impl",
+    srcs = [
+        "listener_manager_impl.cc",
+    ],
+    hdrs = [
+        "listener_impl.h",
+        "listener_manager_impl.h",
+    ],
     deps = [
         ":configuration_lib",
         ":drain_manager_lib",
@@ -272,12 +308,9 @@ envoy_cc_library(
         "//include/envoy/server:listener_manager_interface",
         "//include/envoy/server:transport_socket_config_interface",
         "//include/envoy/server:worker_interface",
-        "//source/common/api:os_sys_calls_lib",
         "//source/common/config:utility_lib",
         "//source/common/init:manager_lib",
-        "//source/common/network:connection_balancer_lib",
         "//source/common/network:listen_socket_lib",
-        "//source/common/network:resolver_lib",
         "//source/common/network:socket_option_factory_lib",
         "//source/common/network:utility_lib",
         "//source/common/protobuf:utility_lib",
@@ -344,7 +377,7 @@ envoy_cc_library(
         ":connection_handler_lib",
         ":guarddog_lib",
         ":listener_hooks_lib",
-        ":listener_manager_lib",
+        ":listener_lib",
         ":ssl_context_manager_lib",
         ":worker_lib",
         "//include/envoy/event:dispatcher_interface",
